+++ src/services/inplayer.subscription.service.ts
@@ -1,7 +1,7 @@
 import i18next from 'i18next';
 import InPlayer, { PurchaseDetails, Card, GetItemAccessV1, SubscriptionDetails as InplayerSubscription } from '@inplayer-org/inplayer.js';
 
-import type { PaymentDetail, Subscription, Transaction, UpdateSubscription } from '#types/subscription';
+import type { PaymentDetail, Subscription, Transaction, UpdateCardDetails, UpdateSubscription } from '#types/subscription';
 import type { Config } from '#types/Config';
 import type { InPlayerError } from '#types/inplayer';
 
@@ -53,21 +53,25 @@ export async function getAllTransactions() {
     throw new Error('Failed to get transactions');
   }
 }
-
 export async function getActivePayment() {
   try {
     const { data } = await InPlayer.Payment.getDefaultCreditCard();
     const cards: PaymentDetail[] = [];
     for (const currency in data?.cards) {
-      cards.push(formatCardDetails(data.cards?.[currency]));
+      cards.push(
+        // @ts-ignore
+        // TODO fix Card type in InPlayer SDK
+        formatCardDetails({
+          ...data.cards?.[currency],
+          currency: currency,
+        }),
+      );
     }
-
     return cards.find((paymentDetails) => paymentDetails.active) || null;
   } catch {
     return null;
   }
 }
-
 export const getSubscriptions = async () => {
   return {
     errors: [],
@@ -90,8 +94,17 @@ export const updateSubscription: UpdateSubscription = async ({ offerId, unsubscr
   }
 };
 
-const formatCardDetails = (card: Card): PaymentDetail => {
-  const { number, exp_month, exp_year, card_name, card_type, account_id } = card;
+export const updateCardDetails: UpdateCardDetails = async ({ cardName, cardNumber, cvc, expMonth, expYear, currency }) => {
+  try {
+    const response = await InPlayer.Payment.setDefaultCreditCard({ cardName, cardNumber, cvc, expMonth, expYear, currency });
+
+    return { responseData: response.data, errors: [] };
+  } catch {
+    throw new Error('Failed to update card details');
+  }
+};
+const formatCardDetails = (card: Card & { card_type: string; account_id: number; currency: string }): PaymentDetail => {
+  const { number, exp_month, exp_year, card_name, card_type, account_id, currency } = card;
   const zeroFillExpMonth = `0${exp_month}`.slice(-2);
   return {
     customerId: account_id.toString(),
@@ -102,6 +115,7 @@ const formatCardDetails = (card: Card): PaymentDetail => {
       cardExpirationDate: `${zeroFillExpMonth}/${exp_year}`,
     },
     active: true,
+    currency,
   } as PaymentDetail;
 };